# Conflicts:
#	ProjectManagementSystemGUI-master/src/model/Project.java
#	ProjectManagementSystemGUI-master/src/model/RequirementList.java

+++ ProjectManagementSystemGUI-master/src/model/RequirementList.java
@@ -92,9 +92,9 @@ public class RequirementList
         getRequirementByID(requirementID).setOrderNum(orderNum);
     }
 
-    public double getRequirementsListTotalHoursOfWork()
+    public int getRequirementsListTotalHoursOfWork()
     {
-        double hours = 0;
+        int hours = 0;
         for (int i = 0; i < requirements.size(); i++)
         {
             hours += requirements.get(i).getSpentHours();
